% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator.R
\name{tcpCompare}
\alias{tcpCompare}
\title{Compares cross-products}
\usage{
tcpCompare(x, y)
}
\arguments{
\item{x}{A matrix-like object.}

\item{y}{Another matrix-like object.}
}
\value{
A logical scalar indicating whether \code{tcrossprod(x, y)} is identical to \code{x \%*\% t(y)}.
}
\description{
Compare the output of \code{tcrossprod} with standard matrix multiplication.
}
\examples{
NC <- 10
NP <- 20
NR <- 50

comp <- matrix(rnorm(NC*NP), ncol=NP)
rot <- Matrix::rsparsematrix(NR, NP, 0.1)

tcpCompare(comp, rot)

}
\author{
Aaron Lun
}
